feat: Enable length based blacken substitution
Currently, blacken function replaces the characters with the
replacement without considering that the length can already
be an issue, as this could expose sensible information or
at least can result in that the malicious actor can gain
further information about the data.

Resolve this issue by using a configurable length with fallback and
enabling default substitution.

Considered adding a default length for blacken
that can resolve most cases. If intended blacken = 5 this
would not really help, but anything else are even less useful.

Considered Random and fallback to default length based on
previous data.

Due to the lack of knowledge of previous data, we can not
use previous data to always use X length without rework.

Implemented in both occurrences sapl-lang and sapl-spring-security.
We should strive to unify both string operation in a util module or
similar.